+++ R/expect-setequal.R
@@ -4,11 +4,9 @@
 #'    and that every element of `y` occurs in `x`.
 #' * `expect_contains(x, y)` tests that `x` contains every element of `y`
 #'   (i.e. `y` is a subset of `x`).
-#' * `expect_not_contains(x, y)` tests that `x` contains none of the elements
-#'   of `y` (i.e. `y` is disjoint from `x`).
 #' * `expect_in(x, y)` tests that every element of `x` is in `y`
 #'   (i.e. `x` is a subset of `y`).
-#' * `expect_not_in(x, y)` tests that no element of `x` is in `y`
+#' * `expect_disjoint(x, y)` tests that no element of `x` is in `y`
 #'   (i.e. `x` is disjoint from `y`).
 #' * `expect_mapequal(x, y)` treats lists as if they are mappings between names
 #'   and values. Concretely, this drops `NULL`s in both objects and sorts
@@ -121,32 +119,6 @@ expect_contains <- function(object, expected) {
   pass(act$val)
 }
 
-#' @export
-#' @rdname expect_setequal
-expect_not_contains <- function(object, expected) {
-  act <- quasi_label(enquo(object))
-  exp <- quasi_label(enquo(expected))
-
-  check_vector(object)
-  check_vector(expected)
-
-  exp_found <- exp$val %in% act$val
-  if (any(exp_found)) {
-    msg_exp <- sprintf(
-      "Expected %s to contain none of the values in %s.",
-      act$lab,
-      exp$lab
-    )
-    msg_act <- c(
-      sprintf("Actual: %s", values(act$val)),
-      sprintf("Expected: none of %s", values(exp$val)),
-      sprintf("Found: %s", values(exp$val[exp_found]))
-    )
-    fail(c(msg_exp, msg_act))
-  }
-
-  pass(act$val)
-}
 
 #' @export
 #' @rdname expect_setequal
@@ -177,24 +149,24 @@ expect_in <- function(object, expected) {
 
 #' @export
 #' @rdname expect_setequal
-expect_not_in <- function(object, expected) {
+expect_disjoint <- function(object, expected) {
   act <- quasi_label(enquo(object))
   exp <- quasi_label(enquo(expected))
 
   check_vector(object)
   check_vector(expected)
 
-  act_found <- act$val %in% exp$val
-  if (any(act_found)) {
+  act_common <- act$val %in% exp$val
+  if (any(act_common)) {
     msg_exp <- sprintf(
-      "Expected %s to contain no values from %s.",
+      "Expected %s to be disjoint from %s.",
       act$lab,
       exp$lab
     )
     msg_act <- c(
       sprintf("Actual: %s", values(act$val)),
       sprintf("Expected: none of %s", values(exp$val)),
-      sprintf("Invalid: %s", values(act$val[act_found]))
+      sprintf("Invalid: %s", values(act$val[act_common]))
     )
     fail(c(msg_exp, msg_act))
   }